Refactor plugin system: state management, cleanup, and minor improvements
