+++ src/physics/plugin/execution.rs
@@ -46,9 +46,14 @@ impl ExecutionStrategy for SequentialStrategy {
 }
 
 /// Plugin executor that manages plugin execution
-#[derive(Debug)]
 pub struct PluginExecutor {
     strategy: Box<dyn ExecutionStrategy>,
+}
+
+impl std::fmt::Debug for PluginExecutor {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("PluginExecutor").finish()
+    }
 }
 
 impl PluginExecutor {
