+++ src/physics/plugin/metadata.rs
@@ -33,16 +33,4 @@ pub trait PluginConfig: Debug + Send + Sync {
 }
 
 /// Plugin lifecycle state
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum PluginState {
-    /// Plugin is created but not initialized
-    Created,
-    /// Plugin is initialized and ready to run
-    Initialized,
-    /// Plugin is currently processing
-    Running,
-    /// Plugin encountered an error
-    Error,
-    /// Plugin has been finalized
-    Finalized,
-}
+
