+++ src/solver/pstd/plugin.rs
@@ -229,8 +229,8 @@ impl crate::physics::plugin::Plugin for PstdPlugin {
         Ok(())
     }
 
-    fn diagnostics(&self) -> HashMap<String, f64> {
-        HashMap::new()
+    fn diagnostics(&self) -> String {
+        format!("PSTD Plugin - State: {:?}", self.state)
     }
 
     fn reset(&mut self) -> KwaversResult<()> {
