+++ src/solver/time_integration/mod.rs
@@ -34,7 +34,7 @@ pub use traits::{MultiRateConfig, TimeStepper, TimeStepperConfig, TimeStepperTyp
 
 use crate::error::{KwaversError, ValidationError};
 use crate::grid::Grid;
-use crate::physics::plugin::crate::physics::plugin::Plugin;
+use crate::physics::plugin::Plugin;
 use crate::KwaversResult;
 use ndarray::Array3;
 use std::collections::HashMap;
